src/requirements.txt
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
msgpack==1.0.8
oauth2client==4.1.3
oauthlib==3.2.2
patool==2.2.0
plyvel==1.5.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
PySide6==6.7.1
PySide6_Addons==6.7.1
PySide6_Essentials==6.7.1
ratelimit==2.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
shiboken6==6.7.1
six==1.16.0
tomli==2.0.1
tomlkit==0.12.5
uritemplate==4.1.1
urllib3==2.2.1
